pyproject.toml
[project]
name = "stance-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "apify-client>=2.5.0",
    "clean-text>=0.7.1",
    "datasets>=4.8.3",
    "instaloader>=4.15.1",
    "matplotlib>=3.10.8",
    "nltk>=3.9.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "regex>=2026.2.28",
    "sastrawi>=1.0.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "seqeval>=1.2.2",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]
